Reorganize and split files
