+++ src/service/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(feature = "rtu")]
 pub mod rtu;
+
 #[cfg(feature = "tcp")]
 pub mod tcp;
